cleanup: folder structure change
