--- Auth_Api/src/Auth_Api.WebApi.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35327.3
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Auth_Api.WebApi", "Auth_Api.WebApi\Auth_Api.WebApi.csproj", "{30B9E7CA-049C-431A-B23E-1E5F0CD9A185}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Auth_Api", "Auth_Api\Auth_Api.csproj", "{429C9491-3BC1-4A38-9ED5-C310B002677C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "..\tests\Tests.csproj", "{A889C8EA-4712-4B5F-BC77-139F5600B67E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{30B9E7CA-049C-431A-B23E-1E5F0CD9A185}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{30B9E7CA-049C-431A-B23E-1E5F0CD9A185}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{30B9E7CA-049C-431A-B23E-1E5F0CD9A185}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{30B9E7CA-049C-431A-B23E-1E5F0CD9A185}.Release|Any CPU.Build.0 = Release|Any CPU
-		{429C9491-3BC1-4A38-9ED5-C310B002677C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{429C9491-3BC1-4A38-9ED5-C310B002677C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{429C9491-3BC1-4A38-9ED5-C310B002677C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{429C9491-3BC1-4A38-9ED5-C310B002677C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A889C8EA-4712-4B5F-BC77-139F5600B67E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A889C8EA-4712-4B5F-BC77-139F5600B67E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A889C8EA-4712-4B5F-BC77-139F5600B67E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A889C8EA-4712-4B5F-BC77-139F5600B67E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {37337B0A-8552-4B3E-85D7-D653910E0E42}
-	EndGlobalSection
-EndGlobal
