--- Auth_Api/src/Auth_Api.WebApi/Auth_Api.WebApi.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Version>0.0.1</Version>
-    <UserSecretsId>16861e60-522b-4524-96a3-f302ee1a7e98</UserSecretsId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Auth_Api\Auth_Api.csproj" />
-  </ItemGroup>
-
-</Project>
