--- Auth_Api/src/Auth_Api.WebApi/Program.cs
@@ -1,38 +1,0 @@
-using Auth_Api.Services.JWT;
-using Auth_Api.WebApi.Configurations;
-using Microsoft.Extensions.FileProviders;
-
-var builder = WebApplication.CreateBuilder(args);
-
-builder.Services.AddCustomConfigurations();
-builder.Services.AddJwtConfigurations(builder.Configuration["JwtPrivateKey"]);
-
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-builder.Services.AddHealthChecks();
-
-var app = builder.Build();
-
-app.MapHealthChecks("/health");
-
-// Configure the HTTP request pipeline.
-//if (app.Environment.IsDevelopment())
-//{
-app.UseSwagger();
-app.UseSwaggerUI();
-//}
-
-app.UseDefaultFiles();
-app.UseStaticFiles(new StaticFileOptions()
-{
-    FileProvider = new PhysicalFileProvider(Path.Combine(Directory.GetCurrentDirectory(), @"wwwroot"))
-});
-
-app.UseHttpsRedirection();
-
-app.UseAuthentication();
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.Run();
