--- Auth_Api/src/Auth_Api.WebApi/wwwroot/index.html
@@ -1,9 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <title>PraJa Api</title>
-</head>
-<body style="background-color: black; color: white">
-    <p>API Ok</p>
-</body>
-</html>
