--- Auth_Api/src/Auth_Api/Services/JWT/JwtConfigurations.cs
@@ -1,34 +1,0 @@
-﻿using Microsoft.IdentityModel.Tokens;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using System.Text;
-
-namespace Auth_Api.Services.JWT
-{
-    public static class JwtConfigurations
-    {
-        public static IServiceCollection AddJwtConfigurations(this IServiceCollection services, string secret)
-        {
-            services.AddScoped<TokenService>();
-            services.AddAuthentication(x =>
-            {
-                x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            })
-            .AddJwtBearer(x =>
-            {
-                x.RequireHttpsMetadata = false;
-                x.SaveToken = true;
-                x.TokenValidationParameters = new TokenValidationParameters
-                {
-                    ValidateIssuerSigningKey = true,
-                    IssuerSigningKey = new SymmetricSecurityKey(Encoding.ASCII.GetBytes(secret)),
-                    ValidateIssuer = false,
-                    ValidateAudience = false
-                };
-            });
-
-            return services;
-        }
-    }
-}
