--- Auth_Api/src/Auth_Api/Services/JWT/TokenService.cs
@@ -1,42 +1,0 @@
-﻿using Auth_Api.Users;
-using Microsoft.Extensions.Configuration;
-using Microsoft.IdentityModel.Tokens;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Text;
-
-namespace Auth_Api.Services.JWT;
-
-public class TokenService
-{
-    private readonly IConfiguration _config;
-
-    public TokenService(IConfiguration config)
-    {
-        _config = config;
-    }
-
-    public string GenerateToken(User user)
-    {
-        var tokenHandler = new JwtSecurityTokenHandler();
-        
-        var key = Encoding.ASCII.GetBytes(_config["JwtPrivateKey"]);
-
-        var tokenDescriptor = new SecurityTokenDescriptor
-        {
-            Subject = new ClaimsIdentity(new Claim[]
-            {
-                new Claim(ClaimTypes.Name, user.Name.ToString()),
-                new Claim(ClaimTypes.Role, user.Role.ToString())
-                //new Claim(ClaimTypes.Email, user.Email.ToString())
-            }),
-            
-            Expires = DateTime.UtcNow.AddHours(2),
-
-            SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
-        };
-
-        var token = tokenHandler.CreateToken(tokenDescriptor);
-        return tokenHandler.WriteToken(token);
-    }
-}
