--- Auth_Api/src/Auth_Api/Users/Features/RegisterUser/v1/RegisterUserHandler.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Auth_Api.Users.Features.RegisterUser.v1;
-
-public class RegisterUserHandler
-{
-}
