--- Auth_Api/src/Auth_Api/Users/UserRepository.cs
@@ -1,17 +1,0 @@
-﻿using CSharpFunctionalExtensions;
-
-namespace Auth_Api.Users;
-
-public sealed class UserRepository
-{
-    public Maybe<User> Get(string username, string password)
-    {
-        var users = new List<User>()
-        {
-            User.Create(Guid.NewGuid(), "batman", "batman@email.com", "batman", "manager").Value,
-            User.Create(Guid.NewGuid(), "robin", "robin@email.com", "robin", "employee").Value
-        };
-
-        return users.Where(x => x.Name.ToLower() == username.ToLower() && x.Password == password).FirstOrDefault();
-    }
-}
