--- Auth_Api/templates.config/develop-deploy.yml
@@ -1,40 +1,0 @@
-steps:
-- task: UseDotNet@2
-  displayName: use sdk
-  inputs:
-    packageType: 'sdk'
-    version: '8.x'
-
-- task: DotNetCoreCLI@2
-  displayName: restore
-  inputs:
-    command: 'restore'
-    projects: '$(solution)'
-
-- task: DotNetCoreCLI@2
-  displayName: test
-  inputs:
-    command: 'test'
-    projects: '$(solution)'
-
-- task: DotNetCoreCLI@2
-  displayName: build
-  inputs:
-    command: 'build'
-    projects: '$(solution)'
-    arguments: '--configuration $(buildConfiguration)'
-
-- task: DotNetCoreCLI@2
-  displayName: publish
-  inputs:
-    command: 'publish'
-    publishWebProjects: true
-    arguments: '--configuration $(buildConfiguration) --output $(Build.ArtifactStagingDirectory)'
-    zipAfterPublish: true
-
-- task: PublishBuildArtifacts@1
-  displayName: publish artifacts
-  inputs:
-    pathtoPublish: '$(Build.ArtifactStagingDirectory)'
-    artifactName: 'drop'
-    publishLocation: 'Container'
