--- azure-pipelines.yml
@@ -1,25 +1,0 @@
-trigger:
-- main
-
-parameters:
-  #- name: ambiente
-    #displayName: Ambiente
-    #type: string
-    #default: develop
-    #values:
-      #- develop
-      #- homolog
-  - name: deploy
-    displayName: Publicar?
-    type: boolean
-    default: false
-
-variables:
-  solution: 'Auth_Api/src/Auth_Api.WebApi.sln'
-  buildPlatform: 'Any CPU'
-  buildConfiguration: 'Release'
-
-steps:
-# - template: Auth_Api/templates.config/develop-deploy.yml
-- ${{ if eq(parameters.deploy, true) }}:
-  - template: Auth_Api/templates.config/homolog-deploy.yml
