+++ src/AuthApi.Application/Features/Users/AuthenticateUser/v1/AuthenticateUserCommand.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AuthApi.Application.Features.Users.AuthenticateUser.v1;
+
+public class AuthenticateUserCommand
+{
+}
