+++ src/AuthApi.Application/Features/Users/RegisterUser/v1/RegisterUserCommand.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AuthApi.Application.Features.Users.RegisterUser.v1;
+
+public class RegisterUserCommand
+{
+}
