+++ src/AuthApi.Application/Features/Users/User.cs
@@ -1,0 +1,52 @@
+﻿using AuthApi.Shared.Resource;
+using CSharpFunctionalExtensions;
+
+namespace AuthApi.Application.Features.Users;
+
+public sealed class User
+{
+    public Guid Id { get; private set; }
+    public string Name { get; private set; }
+    public string Email { get; private set; }
+    public string Password { get; private set; }
+    public string Role { get; private set; }
+
+    private User(Guid id, string name, string email, string password, string role)
+    {
+        Id = id;
+        Name = name;
+        Email = email;
+        Password = password;
+        Role = role;
+    }
+
+    public static Result<User> Create(Guid id, string name, string email, string password, string role)
+    {
+        //if (Guid.TryParse(id, out resultId))
+        //{
+        //    return Result.Failure<User>(AuthApi_Resource.NAME_REQUIRED);
+        //}
+
+        if (string.IsNullOrEmpty(name))
+        {
+            return Result.Failure<User>(AuthApi_Resource.NAME_REQUIRED);
+        }
+
+        if (string.IsNullOrEmpty(email))
+        {
+            return Result.Failure<User>(AuthApi_Resource.EMAIL_REQUIRED);
+        }
+
+        if (string.IsNullOrEmpty(password))
+        {
+            return Result.Failure<User>(AuthApi_Resource.PASSWORD_REQUIRED);
+        }
+
+        if (string.IsNullOrEmpty(role))
+        {
+            return Result.Failure<User>("Role é obrigatória.");
+        }
+
+        return Result.Success(new User(id, name, email, password, role));
+    }
+}
