+++ src/AuthApi.Shared/Security/JWT/JwtConfigurations.cs
@@ -1,0 +1,33 @@
+﻿using Microsoft.IdentityModel.Tokens;
+using Microsoft.Extensions.DependencyInjection;
+using System.Text;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+
+namespace AuthApi.Shared.Security.JWT;
+
+public static class JwtConfigurations
+{
+    public static IServiceCollection AddJwtConfigurations(this IServiceCollection services, string secret)
+    {
+        services.AddScoped<TokenService>();
+        services.AddAuthentication(x =>
+        {
+            x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+            x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+        })
+        .AddJwtBearer(x =>
+        {
+            x.RequireHttpsMetadata = false;
+            x.SaveToken = true;
+            x.TokenValidationParameters = new TokenValidationParameters
+            {
+                ValidateIssuerSigningKey = true,
+                IssuerSigningKey = new SymmetricSecurityKey(Encoding.ASCII.GetBytes(secret)),
+                ValidateIssuer = false,
+                ValidateAudience = false
+            };
+        });
+
+        return services;
+    }
+}
