+++ src/AuthApi.WebApi/wwwroot/index.html
@@ -1,0 +1,9 @@
+﻿<!DOCTYPE html>
+<html>
+<head>
+    <title>Auth Api</title>
+</head>
+<body style="background-color: black; color: white">
+    <p>API Ok</p>
+</body>
+</html>
